Changes to get lightbox2 working (still not working)
In lightbox2 folder, moved JS and CSS to own folders so images for
lightbox could be found. Now I get an error when clicking images.

Changed unique “data-lightbox” attribute value to “photos” for all
anchor elements in gallery.

Getting new error when clicking on image:
Uncaught TypeError: this.$lightbox.css(...).fadeIn is not a function
    at Lightbox.start (lightbox.js:248)
    at HTMLAnchorElement.<anonymous> (lightbox.js:87)
    at HTMLBodyElement.dispatch (jquery-3.2.0.slim.min.js:3)
    at HTMLBodyElement.q.handle (jquery-3.2.0.slim.min.js:3)